moved common logic in the separate IEmailVerifier.cs
