+++ Application/Abstractions/EmailService/IEmailVerifier.cs
@@ -1,0 +1,8 @@
+using Application.Responses;
+
+namespace Application.Abstractions.EmailService;
+
+public interface IEmailVerifier
+{
+    Task<VerifyEmailResult> VerifyEmailAsync(string token, CancellationToken cancellationToken);
+}
