+++ Application/Abstractions/Security/IVerificationService.cs
@@ -6,7 +6,5 @@ public interface IVerificationService
 {
     Task<SendVerificationLinkResult> SendVerificationLinkAsync(string email, bool isEmailVerification,
                                                             CancellationToken cancellationToken);
-    //todo move to the another service
-    Task<VerifyEmailResult> VerifyEmailAsync(string token, CancellationToken cancellationToken);
     
 }
