+++ Infrastructure/InfrastructureDependencyInjection.cs
@@ -15,6 +15,7 @@ public static class InfrastructureDependencyInjection
     {
         services.AddScoped<IAuthHelper, AuthHelper.AuthHelper>();
         services.AddScoped<IEmailService, EmailService>();
+        services.AddScoped<IEmailVerifier, EmailVerifier>();
         services.AddScoped<IJwtGenerator, JwtGenerator>();
         services.AddScoped<IPasswordHasher, PasswordHasher>();
         services.AddScoped<ITokenGenerator, TokenGenerator>();
